"fixed" dossiers - workaround possible gremlin bug
This is a very strange problem and I'm not certain exactly what to do
about it.  Basically, when creating dossiers on individual users it
appears that there are times when the pipe gets emptied out before all
of the data have been flushed through the pipe. I switched the
implementation some to use a loop for each type of EdgeType. It seems to
work.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>