Script to create a profile for each project
I know Corey was supposed to do this, but I couldn't get the task list
while I was on the plane. This goes about creating a dossier of sorts
for projects.

It expects that you have a directory called rails.data which contains
the file rails.db.20120505.projects.txt

From there it will generate a useful dossier for each project in the
list.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>