Created script to return all repo users
This script gets all of the users involved with each project in the various
different classes as provided by the methods in Helpers.groovy. These can
then be used to check the overlap between different projects.

In the process of writing this it was necessary to refactor a few pieces of
code, which resulted in the creation of RailsHelpers.groovy.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>