Corrected count of users associated with repos
This resulted from the switch away from the original embedded germlin
code to the new java pipes code. I wasn't creating a new pipe for
each pipe and therefore was getting garbage results and an obscene
number of users for most projects. This appears to have fixed it.

I should stress, this DID NOT affect out results in an meaningful
way. The queries for the analysis are all written in groovy and were
not impacted by these changes. At some point, I'll need to make it so
all the queries are done from a single source.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>