Fixed NullPointerException in processing events
For some reason in extremely rare situations there can be events that
have an actor but don't actually have any information about the user.
In these cases the code for importing events was crashing. I added
a check to see not only if the actor was not null, but if the user
contained in the actor was not null too. This appears to fix the
problem.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>