Removed logback.xml from packed jar files
logback.xml was packaged inside of the jar file, which made gitminer
very unhappy when used as an embedded component that also required
logback.xml. Now it copies logback.xml to the target directory. Either
copy it to the base directory if you want color or use the -l flag
on the command line. For example:

    ./github.sh -c configuration.properties -l target/logback.xml

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>