Updated to GovSciGraph 2.1.0. Renamed to GitMiner
This is a possibly breaking change, but it appears it works. I've
updated the code to use govscigraph 2.1.0 which tracks the 2.1.0
series of TinkerPop. This required some changes to various parts of
the code to ensure that we were using the updated package names.

The program was also renamed to gitminer.

A fix was applied for NullPointerExceptions at the end of execution.

Closes #32.

Signed-off-by: Patrick Wagstrom <patrick@wagstrom.net>