Updated to address GitHub URL changes
GitHub recently changed URLs to require https instead of http, this
fixes the code so we can pull repositories.

Also, added in some defaults for some of the properties within
repository loader. In the future the property code should be merged
with the property code from the main part of gitminer.

Signed-off-by: Patrick Wagstrom <pwagstro@us.ibm.com>